requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.12.4,<4.0
djangorestframework-simplejwt>=5.0.0,<6.0
mysqlclient>=2.0.3,<3.0
django-cors-headers>=3.5.0,<4.0
django-environ>=0.4.5,<1.0
drf-yasg>=1.20.0,<2.0
